refactor: move completion code into internal/completion package
Shell-completion helpers lived as sibling files of main.go in package
main. Relocate them to internal/completion to match the project layout
(thin main.go, packages under internal/), exporting RegisterShellCompletions
and NewCompletionCommand for main.go to consume.
