+++ internal/completion/completion_test.go
@@ -1,4 +1,4 @@
-package main
+package completion
 
 import (
 	"os"
